Separation of Back-end and Front-end. Editor and Runtime Console in sync
It was refactor a lot of the code so that it was possible to run the console in runtime and in the editor and still have them in sync. Both Editor and Runtime Console will have their own front-end while using the same back-end for input handling, submiting commands, autotabbing and the rest.